+++ src/mako/http/response/senders/ResponseSenderInterface.php
@@ -23,5 +23,5 @@ interface ResponseSenderInterface
 	 * @param \mako\http\Request  $request  Request instance
 	 * @param \mako\http\Response $response Response instance
 	 */
-	public function send(Request $request, Response $response);
+	public function send(Request $request, Response $response): void;
 }